+++ Hosting/processor_holder.cpp
@@ -354,7 +354,7 @@ void gmpi_processor::setPresetUnsafe(std::string& chunk)
 			formatVersion = 1;
 			try
 			{
-				fourCC = std::stoul(hexcode, nullptr, 16);
+				fourCC = static_cast<int32_t>(std::stoul(hexcode, nullptr, 16));
 			}
 			catch (...)
 			{
